+++ test/libsyncengine/integration/testintegration.h
@@ -141,14 +141,14 @@ class TestIntegration : public CppUnit::TestFixture, public TestBase {
         // Resolves a possibly multi-segment relative path (e.g. "A/AA/BBB") by walking down the remote tree one
         // path component at a time, starting from rootParentId. Returns an invalid RemoteFileInfo as soon as any
         // segment along the way cannot be found (in particular if the leaf itself doesn't exist).
-        RemoteFileInfo getRemoteFileInfoByPath(int driveDbId, const NodeId &rootParentId, const SyncPath &relativePath) const;
-        int64_t countItemsInRemoteDir(int driveDbId, const NodeId &parentId) const;
+        RemoteFileInfo getRemoteFileInfoByPath(int64_t driveDbId, const NodeId &rootParentId, const SyncPath &relativePath) const;
+        int64_t countItemsInRemoteDir(int64_t driveDbId, const NodeId &parentId) const;
 
         log4cplus::Logger _logger;
         std::shared_ptr<SyncPal> _syncPal = nullptr;
         std::shared_ptr<ParmsDb> _parmsDb = nullptr;
 
-        int _driveDbId = 0;
+        int64_t _driveDbId = 0;
         LocalTemporaryDirectory _localSyncDir;
         RemoteTemporaryDirectory _remoteSyncDir{"testIntegration"};
         LocalTemporaryDirectory _localTempDir{"testIntegration"};
